requirements.txt
psycopg2-binary==2.9.9
pykakasi==2.2.1
googletrans==4.0.0-rc1